# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet Operations API
  version: 1.0.0
  description: 车队经营管理系统统一 API 契约。详细业务规则以项目需求基线文档为准。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /routes:
    get:
      tags: [Routes]
      summary: 线路列表
      x-permission: route:list
      responses: {'200': {description: OK}}
    post:
      tags: [Routes]
      summary: 创建线路
      x-permission: route:create
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RouteCreateRequest'}}}}
      responses: {'200': {description: OK}}
  /routes/{id}:
    get:
      tags: [Routes]
      summary: 线路详情
      x-permission: route:list
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    put:
      tags: [Routes]
      summary: 发车前编辑线路
      x-permission: route:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    delete:
      tags: [Routes]
      summary: 发车前逻辑删除
      x-permission: route:delete
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /routes/{id}/publish:
    post:
      tags: [Routes]
      summary: 发布线路并强校验司机车辆
      x-permission: route:publish
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /routes/{id}/depart:
    post:
      tags: [Routes]
      summary: 发车并冻结司机车辆车挂快照
      x-permission: route:depart
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /routes/{id}/unload:
    post:
      tags: [Routes]
      summary: 提交毛重皮重并卸货
      x-permission: route:unload
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UnloadRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /routes/{id}/cancel:
    post:
      tags: [Routes]
      summary: 发车前取消
      x-permission: route:cancel
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /routes/{id}/void:
    post:
      tags: [Routes]
      summary: 运输中作废
      x-permission: route:void
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /routes/{id}/reactivate:
    post:
      tags: [Routes]
      summary: 作废线路重新启用并强校验
      x-permission: route:reactivate
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /routes/{id}/weight-adjustments:
    post:
      tags: [Routes]
      summary: 财务调整毛重皮重
      x-permission: route:weight:adjust
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WeightAdjustmentRequest'}}}}
      responses: {'200': {description: OK}}
  /routes/{id}/weight-versions:
    get:
      tags: [Routes]
      summary: 查询线路重量版本历史
      description: 返回司机原始重量和每次财务纠错版本，包含修改前后可比对字段、修改人、时间、原因和 WEIGHT_ADJUST 附件 ID。
      x-permission: route:list
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiResponse'}
                  - type: object
                    properties:
                      data:
                        type: array
                        items: {$ref: '#/components/schemas/RouteWeightVersion'}
...
  /settlements/{yearMonth}/versions:
    get:
      tags: [Settlements]
      summary: 月度结算版本列表
      x-permission: settlement:view
      parameters: [{$ref: '#/components/parameters/YearMonth'}]
      responses: {'200': {description: OK}}
    post:
      tags: [Settlements]
      summary: 生成新的不可变结算版本
      description: 按业务发生日期生成不可变快照，保存收入、线路费用、日常费用、司机工资、利润，并相对上一版本生成 ADD、DELETE、MODIFY、REASSIGN、WEIGHT_ADJUST、REVERSAL 差异。
      x-permission: settlement:generate
      parameters: [{$ref: '#/components/parameters/YearMonth'}]
      responses: {'200': {description: OK}}
  /settlements/versions/{id}/details:
    get:
      tags: [Settlements]
      summary: 结算快照明细
      x-permission: settlement:view
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /settlements/versions/{id}/diff:
    get:
      tags: [Settlements]
      summary: 相对上一版本差异
      description: 差异包含修改前后快照、收入/支出/利润影响、真实业务操作人、操作时间和原因；与结算版本生成人区分。
      x-permission: settlement:diff:view
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
    YearMonth: {name: yearMonth, in: path, required: true, schema: {type: string, pattern: '^\\d{4}-\\d{2}$'}}
  responses:
    BusinessConflict:
      description: 业务规则冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiResponse'}
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string, examples: ['0']}
        message: {type: string}
        data: {}
        requestId: {type: string}
...
    DecimalWeight:
      type: string
      pattern: '^\\d{1,15}(\\.\\d{1,3})?$'
      description: 重量/电量/气量 DECIMAL(18,3)，接口以字符串传输避免浮点误差。
    DecimalUnitPrice:
      type: string
      pattern: '^\\d{1,14}(\\.\\d{1,4})?$'
      description: 单价 DECIMAL(18,4)，接口以字符串传输避免浮点误差。
...
    RouteCreateRequest:
      type: object
      required: [businessDate, customerId, productId, direction, loadingPlace, unloadingPlace, taxUnitPrice]
      properties:
        businessDate: {type: string, format: date}
        customerId: {type: integer, format: int64}
        productId: {type: integer, format: int64}
        direction: {type: string, enum: [OUTBOUND, RETURN]}
        loadingPlace: {type: string}
        unloadingPlace: {type: string}
        assignedDriverId: {type: [integer, 'null'], format: int64}
        taxUnitPrice: {$ref: '#/components/schemas/DecimalUnitPrice'}
        externalRouteNo: {type: [string, 'null']}
        tripSequence: {type: [string, 'null']}
    UnloadRequest:
      type: object
      required: [grossWeight, tareWeight]
      properties:
        grossWeight: {$ref: '#/components/schemas/DecimalWeight'}
        tareWeight: {$ref: '#/components/schemas/DecimalWeight'}
    WeightAdjustmentRequest:
      type: object
      required: [grossWeight, tareWeight, reason]
      properties:
        grossWeight: {$ref: '#/components/schemas/DecimalWeight'}
        tareWeight: {$ref: '#/components/schemas/DecimalWeight'}
        reason: {type: string, minLength: 1}
        attachmentIds: {type: array, items: {type: integer, format: int64}}
    RouteWeightVersion:
      type: object
      properties:
        id: {type: integer, format: int64}
        routeId: {type: integer, format: int64}
        versionNo: {type: integer}
        grossWeight: {$ref: '#/components/schemas/DecimalWeight'}
        tareWeight: {$ref: '#/components/schemas/DecimalWeight'}
        netWeight: {$ref: '#/components/schemas/DecimalWeight'}
        loadStandardThreshold: {$ref: '#/components/schemas/DecimalWeight'}
        loadStandardMet: {type: boolean}
        sourceType: {type: string, enum: [UNLOAD, ADJUSTMENT]}
        operatorId: {type: integer, format: int64}
        operationTime: {type: string, format: date-time}
        reason: {type: [string, 'null']}
        attachmentIds: {type: array, items: {type: integer, format: int64}}
...